gpuav: Remove gpu_state_tracker.cpp
