+++ layers/CMakeLists.txt
@@ -239,8 +239,6 @@ target_sources(vvl PRIVATE
     gpu/descriptor_validation/gpuav_image_layout.h
     gpu/descriptor_validation/gpuav_image_layout.cpp
     gpu/core/gpuav.h
-    gpu/core/gpu_state_tracker.cpp
-    gpu/core/gpu_state_tracker.h
     gpu/core/gpuav_record.cpp
     gpu/core/gpuav_setup.cpp
     gpu/debug_printf/debug_printf.cpp
