+++ layers/gpu/instrumentation/gpu_shader_instrumentor.cpp
@@ -17,7 +17,6 @@
 
 #include "gpu/instrumentation/gpu_shader_instrumentor.h"
 
-#include "gpu/core/gpu_state_tracker.h"
 #include "gpu/spirv/module.h"
 #include "chassis/chassis_modification_state.h"
 #include "gpu/shaders/gpu_error_codes.h"
