+++ layers/gpu/resources/gpuav_subclasses.h
@@ -20,7 +20,6 @@
 #include <vector>
 
 #include "external/inplace_function.h"
-#include "gpu/core/gpu_state_tracker.h"
 #include "gpu/descriptor_validation/gpuav_descriptor_set.h"
 #include "gpu/resources/gpu_resources.h"
 
@@ -29,8 +28,48 @@
 #include "state_tracker/buffer_state.h"
 #include "state_tracker/image_state.h"
 #include "state_tracker/cmd_buffer_state.h"
+#include "state_tracker/queue_state.h"
 #include "state_tracker/sampler_state.h"
 #include "state_tracker/ray_tracing_state.h"
+
+namespace gpu {
+class GpuShaderInstrumentor;
+}
+
+namespace gpu_tracker {
+
+class Queue : public vvl::Queue {
+  public:
+    Queue(gpu::GpuShaderInstrumentor &shader_instrumentor_, VkQueue q, uint32_t family_index, uint32_t queue_index,
+          VkDeviceQueueCreateFlags flags, const VkQueueFamilyProperties &queueFamilyProperties, bool timeline_khr);
+    virtual ~Queue();
+
+  protected:
+    vvl::PreSubmitResult PreSubmit(std::vector<vvl::QueueSubmission> &&submissions) override;
+    void PostSubmit(vvl::QueueSubmission &) override;
+    void SubmitBarrier(const Location &loc, uint64_t seq);
+    void Retire(vvl::QueueSubmission &) override;
+
+    gpu::GpuShaderInstrumentor &shader_instrumentor_;
+    VkCommandPool barrier_command_pool_{VK_NULL_HANDLE};
+    VkCommandBuffer barrier_command_buffer_{VK_NULL_HANDLE};
+    VkSemaphore barrier_sem_{VK_NULL_HANDLE};
+    std::deque<std::vector<std::shared_ptr<vvl::CommandBuffer>>> retiring_;
+    const bool timeline_khr_;
+};
+
+class CommandBuffer : public vvl::CommandBuffer {
+  public:
+    CommandBuffer(gpu::GpuShaderInstrumentor &shader_instrumentor_, VkCommandBuffer handle,
+                  const VkCommandBufferAllocateInfo *pCreateInfo, const vvl::CommandPool *pool);
+
+    virtual bool PreProcess(const Location &loc) = 0;
+    virtual void PostProcess(VkQueue queue, const Location &loc) = 0;
+};
+}  // namespace gpu_tracker
+
+VALSTATETRACK_DERIVED_STATE_OBJECT(VkQueue, gpu_tracker::Queue, vvl::Queue)
+VALSTATETRACK_DERIVED_STATE_OBJECT(VkCommandBuffer, gpu_tracker::CommandBuffer, vvl::CommandBuffer)
 
 namespace gpuav {
